+++ apps/vs-agent/src/events/MessageEvents.ts
@@ -1,4 +1,4 @@
-import type { ServerConfig } from '../utils/ServerConfig'
+import type { ServerConfig } from '../utils'
 import type { MessageReceiptsReceivedEvent, MessageState } from '@2060.io/credo-ts-didcomm-receipts'
 import type { DidCommAgentModules } from '@verana-labs/vs-agent-sdk'
